requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
avalanche-lib==0.1.0
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
configparser==5.2.0
cycler==0.11.0
docker-pycreds==0.4.0
faiss-cpu==1.7.2
filelock==3.4.2
fonttools==4.29.0
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gdown==4.2.0
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.5.0
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.43.0
huggingface-hub==0.5.1
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.10.1
joblib==1.1.0
jsonlines==3.0.0
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.1
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.3
pandas==1.4.2
pathtools==0.1.2
Pillow==9.0.0
promise==2.3
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyDeprecate==0.3.1
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.6.0
pytorchcv==0.0.67
pytz==2022.1
PyYAML==6.0
quadprog==0.1.11
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.5.4
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
subprocess32==3.5.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.11.6
torch==1.10.1
torchaudio==0.10.1
torchmetrics==0.7.0
torchvision==0.11.2
tqdm==4.62.3
transformers==4.18.0
typing-extensions==4.0.1
urllib3==1.26.8
wandb==0.12.10
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
